Clarifying the intent of the graph
The graph shown in the book is confusing because of the context.

The text right before the image was speaking about algorithmic complexities. So a person can easily be confused thinking the graph shows the largest data sets that can be handled for a given computational time.

I added clarification that the graph shows the development of computer technology over calendar time.